Refactor replication manager system tests using fixtures.
