+++ s3/replication/manager/tests/system/fixtures/subscribe.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python3
+
+#
+# Copyright (c) 2021 Seagate Technology LLC and/or its Affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# For any questions about this software or licensing,
+# please email opensource@seagate.com or cortx-questions@seagate.com.
+#
+
+import pytest
+
+from s3replicationcommon.templates import subscribe_payload_template
+
+
+@pytest.fixture()
+def subscriber_record(logger):
+    """Prepare subscriber payload for tests."""
+    subscriber_payload = subscribe_payload_template()
+
+    subscriber_payload.pop("id")  # replication manager will generate.
+    subscriber_payload["endpoint"] = "http://localhost:8081/"
+    subscriber_payload["prefetch_count"] = "5"
+
+    return subscriber_payload
